api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
httpx==0.28.1
apscheduler==3.10.4
pydantic==2.10.6
pydantic-settings==2.7.1
PyYAML==6.0.2
openai==1.68.2
pytest==8.3.4
pytest-asyncio==0.25.3
